Simple test for nested lists
Adds a simple test for testing data with nested lists.
The test adds the cities/books/libraries database to the cache,
and adds first a city, then a library. Next, a subscription is added,
before adding a book to the cache, and it is verified that the book
is added to the library.

Signed-off-by: Jonas Jacobsen <joklost@gmail.com>