fix: 🐛 do not delay app start by 2 schema pulls
Because of the sync nature of Houdini Watch Schema plugin start it
was doing pull, afterwards one synced sleep, then another pull and
sync sleep and a then finally it switched to being async and
application could start. In case of configured `schemaPullInterval`
for example to 60 seconds, it would take =~ 120 seconds for an
application to start.